feat(core): add migration script to create explorer-studio app

See merge request dev/leav/leav!1678

+++ apps/core/src/domain/application/applicationDomain.ts
@@ -233,6 +233,10 @@ export default function ({
                 throw new ValidationError({id: {msg: Errors.UNKNOWN_APPLICATION, vars: {application: id}}});
             }
 
+            if (apps.list[0].system) {
+                throw new ValidationError({id: Errors.SYSTEM_APPLICATION_DELETION});
+            }
+
             const deletedApp = await applicationRepo.deleteApplication({id, ctx});
 
             await _sendAppEvent(
